refactor: move the parts components
